pyproject.toml
[project]
name = "flask-configuration"
version = "0.0.1"
description = ""
authors = [
    {name = "Saw Kyaw Myint",email = "sawkyawmyint@metateammyanmar.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"  
dependencies = [
    "alembic (>=1.17.2)",
    "annotated-doc (>=0.0.4)",
    "annotated-types (>=0.7.0)",
    "anyio (>=4.12.1)",
    "blinker (>=1.9.0)",
    "build (>=1.3.0)",
    "cachecontrol (>=0.14.4)",
    "certifi (>=2026.1.4)",
    "charset-normalizer (>=3.4.4)",
    "cleo (>=2.1.0)",
    "click (>=8.3.1)",
    "colorama (>=0.4.6)",
    "crashtest (>=0.4.1)",
    "deprecated (>=1.3.1)",
    "distlib (>=0.4.0)",
    "dnspython (>=2.8.0)",
    "dulwich (>=0.24.10)",
    "email-validator (>=2.3.0)",
    "fastapi (>=0.128.0)",
    "fastapi-cli (>=0.0.20)",
    "fastapi-cloud-cli (>=0.8.0)",
    "fastar (>=0.8.0)",
    "fastjsonschema (>=2.21.2)",
    "filelock (>=3.20.2)",
    "findpython (>=0.7.1)",
    "flask (>=2.3.2)",
    "flask-cors (>=6.0.2)",
    "flask-jwt-extended (>=4.7.1)",
    "flask-limiter (>=4.1.1)",
    "flask-marshmallow (>=1.3.0)",
    "flask-migrate (>=4.1.0)",
    "flask-mysql (>=1.5.2)",
    "flask-pydantic (>=0.14.0)",
    "flask-sqlalchemy (>=3.1.1)",
    "greenlet (>=3.3.0)",
    "h11 (>=0.16.0)",
    "httpcore (>=1.0.9)",
    "httptools (>=0.7.1)",
    "httpx (>=0.28.1)",
    "idna (>=3.11)",
    "installer (>=0.7.0)",
    "itsdangerous (>=2.2.0)",
    "jaraco-classes (>=3.4.0)",
    "jaraco-context (>=6.0.2)",
    "jaraco-functools (>=4.4.0)",
    "jinja2 (>=3.1.6)",
    "keyring (>=25.7.0)",
    "limits (>=5.6.0)",
    "mako (>=1.3.10)",
    "markdown-it-py (>=4.0.0)",
    "markupsafe (>=3.0.3)",
    "marshmallow (>=4.2.0)",
    "marshmallow-sqlalchemy (>=1.4.2)",
    "mdurl (>=0.1.2)",
    "more-itertools (>=10.8.0)",
    "msgpack (>=1.1.2)",
    "ordered-set (>=4.1.0)",
    "packaging (>=25.0)",
    "pbs-installer (>=2025.12.17)",
    "pkginfo (>=1.12.1.2)",
    "platformdirs (>=4.5.1)",
    "poetry (>=2.2.1)",
    "poetry-core (>=2.2.1)",
    "pydantic (>=2.12.5)",
    "pydantic-extra-types (>=2.11.0)",
    "pydantic-settings (>=2.12.0)",
    "pydantic-core (>=2.41.5)",
    "pygments (>=2.19.2)",
    "pyjwt (>=2.10.1)",
    "pymysql (>=1.1.2)",
    "pyproject-hooks (>=1.2.0)",
    "python-dotenv (>=1.2.1)",
    "python-multipart (>=0.0.21)",
    "pywin32-ctypes (>=0.2.3)",
    "pyyaml (>=6.0.3)",
    "rapidfuzz (>=3.14.3)",
    "requests (>=2.32.5)",
    "requests-toolbelt (>=1.0.0)",
    "rich (>=14.2.0)",
    "rich-toolkit (>=0.17.1)",
    "rignore (>=0.7.6)",
    "sentry-sdk (>=2.48.0)",
    "shellingham (>=1.5.4)",
    "sqlalchemy (>=2.0.45)",
    "starlette (>=0.50.0)",
    "timeago (>=1.0.15)",
    "tomlkit (>=0.13.3)",
    "trove-classifiers (>=2025.12.1.14)",
    "typer (>=0.21.1)",
    "typing-inspection (>=0.4.2)",
    "typing-extensions (>=4.15.0)",
    "urllib3 (>=2.6.3)",
    "uvicorn (>=0.40.0)",
    "virtualenv (>=20.36.0)",
    "watchfiles (>=1.1.1)",
    "websockets (>=15.0.1)",
    "werkzeug (>=3.1.4)",
    "wrapt (>=2.0.1)",
    "zstandard (>=0.25.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-socketio (>=5.6.0,<6.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "faker (>=40.1.0,<41.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
